Implement PID-based MMU management.
On a chip the MMU would be reachable via control messages on the
delegation network. In our system the calls
IMemoryAdmin::Reserve/Unreserve provide a pseudo-interface to control
the MMU "remotely" from any core.

So far the MMU would only understand requests of the form "map
physical memory at that virtual address" or "unmap physical memory at
that virtual address". This patch extends the interface so that map is
"on behalf of a given PID". Also a new request is introduced "unmap
all physical memory belonging to the following PID".