Distinguish between suspend and abort in the simulation step function.
The original implementation supported only complete abort of the
simulation. A while ago breakpoints ("bp ..." on the interactive
prompt) and program-driven simulation exceptions have been introduced,
but were implemented using the original abort mechanism, which
prevented any form of resumption. A later patch made all aborts
resumable, which was incorrect for anything but breakpoints and
program-driven exceptions. This patch addresses the situation by
providing separate support for aborts and "resumable exceptions". The
latter is used for breakpoints and program-driven exceptions, while
the former is used for simulaiton errors.