Ensure that exclusive families get created on only 1 core.
Until this point an exclusive allocate (allocate/x) would only reserve
an exclusive context on the 1st target core, then propagate a "normal"
allocation request to the remainder of the place. This patch changes
this behavior so that "allocate/x" entails the flag "single", ie
propagate the place ID but allocate on only 1 core.