Change the memory interface to use data+mask for writes.
Until this point the memory system accepted read/write requests at any
address, including inside cache lines. This patch modifies the
interface so that all memory accesses use addresses aligned on cache
line boundaries, with the size of a full cache line. For
writes/stores/updates, a bitmask is introduced to indicate which bytes
to store.