Make store acknowledgements in the memory system agnostic of client resources.
So far the interface to the memory system was tailored to the
assumption that individual threads are "clients" to memory stores,
that is, only threads can wait on store completions via barriers. To
support this, at each store the TID was sent alongside the request,
with the expectation that the corresponding store acknowledgement
would carry the same TID back to the core.

This patch keeps the logic in the memory system, but using a new type
WClientID instead of TID as the payload for store
acknowledgements. This constitutes the foundation for alternate
consistency protocols where not only thread can be clients of
stores. We foresee either family-based or core-based barriers to be
options.