Read stderr when help2man uses readtrace --version.
For platforms that have help2man so that man files can be created for
the binaries based on their help output, the 'make' requires that
these are generated (and they're useful!). However, the python script
readtrace uses an argument parser to read command-line parameters,
and it favors the newer argparse (2.7) over the deprecated optparse.
This argparse module outputs the response to --version to stderr,
and the Makefile that uses help2man to build readtrace's manpage did
not account for this.

The --no-discard-stderr option for help2man requires version 1.37.1.