Improve the breakpointing interface.
Separate breakpoints on fetch ('bp add f') from breakpoints on execute
('bp add x').  The previous behavior would only break on fetch, which
would include instructions fetched but later flushed before they are
executed, eg. on the delay slot of branches. The new behavior
introduced by this patch is desirable because most of the time users
are interested to break on instructions actually executed. We keep the
option to break on flush since this feature is still desirable to
troubleshoot the pipeline code.