Fix support for bundle creations.
Prior to this patch, if the synchronous form is used (w/ an output
register in the parent thread), there will be *two* register
writebacks from the target place: one for allocation, and one
afterwards for creation.  This is a race condition which could entail
invalid register states in the parent. There should be only one
notification.

Moreover, since the parent thread may issue a sync right away, the
notification should only happen in the create phase, where the
family context is primed for a subsequent sync.