Enable config keys with no value.
Previous to this patch the config file reader was able to associate a
multi-line value to a configuration key. This little-known feature had
never been used. This patch removes this feature (newline characters
terminate the value instead) and simultaneously enables keys that have
no value, eg:

   CmdLineLoads =   # left empty

causes key to have an empty value associated. This way we can create
lists where one of the elements merges to empty, eg:

   SMC:RegInits = a,b,c,$CmdLineLoads
   CmdLineLoads =

This way, if there is no override for CmdLineLoads, reading the value
for SMC:RegInits still works and does not gobble up the configuration
key on the line after CmdLineLoads.