Refactor the performance counter code.
Until this point the performance counters were implemented as a big
"switch" on the address requested by the software. This worked however
the number of counters was defined separately from the switch with a
preprocessor macro. Unfortunately, the macro with the number of
counters easily went out of sync with the number of branches in the
"switch". This patch replaces this mechanism with a vector of
callbacks, and uses the size of the vector as number of counters.

NB: this patch also inverts the two counters "created families" and
"created threads". Beware!