Fix and simplify the slrt for "common" tests.
The slrt code (entry point for SL programs compiled from source code)
contained instructions to preset registers to zero and to print a
message after the test program terminates. The former is not any more
necessary since local registers now automatically read zero. The
latter is not necessary since MGSim now provides a facility for
programs to request termination with a status code.