Register storage trace sets as end of access chain in Memory::RegisterClient.
Memory::RegisterClient populates a storage trace set with all possible
traces through the memory system from a client back to it again. This
trace is constructed using the final buffer on the client side as "end
of chain" for all traces in the set.

This patch makes it possible to use multiple different storages in the
final callback to the client, by declaring a storage set instead of just
a storage for the end of the chain.