Simplify the Directory implemenation.
Instead of using sets/associativity use a single dynamic mapping from
line addresses to tokens. The size of the mapping is bounded
by the protocol to never exceed m_maxNumLines, an architectural constant
(number of cache lines below) so this is really a fixed size hardware
component.

The advantage of this generic approach is that it supports heterogeneous
cache sizes.

The logic of managing the mapping is not described in the model and may
be unrealistic timing-wise. However this is not a regression compared to
the previous situation, which was equally unrealistic.