Simplify the RootDirectory implemenation.
Instead of using sets/associativity use a single dynamic mapping from
addresses to lines. The size of the mapping is bounded by the
protocol to never exceed m_maxNumLines, an architectural
constant (number of cache lines below) so this is really a fixed size
hardware component.

The advantage of this generic approach is that it supports heterogeneous
cache and sub-directory sizes. Also it fixes a bug in the OneLevelCOMA
memory model which had too few entries in the root directory prior to
this patch.

The logic of managing the mapping is not described in the model and may
be unrealistic timing-wise. However this is not a regression compared to
the previous situation, which was equally unrealistic.