Implement a singleton Simulator::Kernel.
Prior to this patch MGSim was structured for full encapsulation,
with the option to run multiple simulators side by side,
each with its own Simulator::Kernel insteance.

This patch introduces support for partial encapsulation,
where a single global Simulator::Kernel instance is
used if the code is compiled with -DSTATIC_KERNEL. This
avoids 1-3 memory lookups on all uses of a Kernel object.
This was tested to improve performance by 5-6% on current
hardware.

The fully encapsulated code is still available (-DSTATIC_KERNEL
not defined), and compiled as separate binaries libmgsim-dyn.a and
mgsim-dyn.