Improve the implementation of Simulator::delegate.
This patch generalizes Simulator::delegate into
Simulator::delegate_gen, to enable its reuse for other
callback signatures. This uses the new variadic template
syntax introduced in C++11. The special case is
still typedef'ed to "delegate" for convenience.

The "GetObject" service is removed, as it is not used
any more in MGSim.