Overhaul the GFX device.
This patch introduces support for "command arrays"
akin to GL display list, so as to accelerate
the composition of rectangles on screen. This is intended
for use to organize the display of multiple benchmarks r
running in parallel on screen.

The documentation is updated accordingly, and an
example "device driver library" is also provided.